backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
django-filter>=23.0
python-dotenv>=1.0
googlemaps>=4.10.0
requests>=2.31.0
